+++ src/js_components/loadMore.js
@@ -1,13 +1,25 @@
 // Копіюємо це!
 // import { loadMore } from './loadMore';
-// import { addLoadMoreBtn } from './addLoadMoreBtn';
 // ...
 // const loadMoreBtn = document.querySelector('.load-more-button');
 // loadMoreBtn.addEventListener('click', () => {
 //     loadMore(функція, що рендерить картки)
 // })
 
-import { addLoadMoreBtn } from "./addLoadMoreBtn";
+const footer = document.querySelector('footer');
+
+export const loadMoreMarkup = '<div class="load-more"><button class="load-more-button" type="button">Load More</button></div> ';
+export function addLoadMoreBtn() {
+    footer.insertAdjacentHTML('beforebegin', loadMoreMarkup);
+}
+
+export function removeLoadMoreBtn() {
+    const buttonContainer = !document.querySelector('.load-more')
+    if (!document.querySelector('.load-more')) {
+        return
+    }
+    buttonContainer.remove()
+}
 
 export let currentPage = 1;
