+++ src/js_components/trendMoviesCards.js
@@ -1,7 +1,6 @@
 import axios from 'axios';
 import { genres } from '../Genres/genres.json';
-import { loadMore } from './loadMore';
-import { addLoadMoreBtn } from './addLoadMoreBtn';
+import { loadMore, addLoadMoreBtn, removeLoadMoreBtn } from './loadMore';
 
 const gallery = document.querySelector('.gallery');
 export const API_KEY = '6308d1a98819d8ffdd4916cbcea5cd95';
@@ -23,6 +22,8 @@ export async function renderTrendMovies(page) {
   }
 }
 
+removeLoadMoreBtn();
+addLoadMoreBtn();
 const loadMoreBtn = document.querySelector('.load-more-button');
 loadMoreBtn.addEventListener('click', () => {
   loadMore(renderTrendMovies);
